refactor: move `tsconfig`-related utils to `utils`
