+++ types/ts-config-option.ts
@@ -1,0 +1,11 @@
+export interface TsConfigOption {
+  /**
+   * Optional filename of the TypeScript config file.
+   */
+  filename?: string
+
+  /**
+   * Root directory where to search for the TypeScript config file.
+   */
+  rootDir: string
+}
